remove longest lable length and custom tooltip
